requirements.txt
gradio>=4.0.0
transformers>=4.36.0
datasets>=2.14.0
torch>=2.0.0
numpy>=1.24.0
evaluate>=0.4.0
Pillow>=9.0.0
opencv-python>=4.7.0
tqdm>=4.64.0
torchvision>=0.15.0
